Moved count logic to ProgressArray to avoid rerenders in Container
