+++ packages/react-insta-stories/src/components/Container.tsx
@@ -8,27 +8,12 @@ import { GlobalCtx } from './../interfaces'
 export default function () {
     const [currentId, setCurrentId] = useState<number>(0)
     const [pause, setPause] = useState<boolean>(true)
-    const [count, setCount] = useState<number>(0)
     const [bufferAction, setBufferAction] = useState<boolean>(true)
     const [videoDuration, setVideoDuration] = useState<number>(0)
 
     let mousedownId = useRef<NodeJS.Timeout>()
-    let animationFrameId = useRef<number>()
 
-    const { width, height, defaultInterval, stories, loop, currentIndex, isPaused, onStoryStart, onStoryEnd, onAllStoriesEnd } = useContext<GlobalCtx>(GlobalContext)
-
-    useEffect(() => {
-        if (!pause) {
-            animationFrameId.current = requestAnimationFrame(incrementCount)
-        }
-        return () => {
-            cancelAnimationFrame(animationFrameId.current)
-        }
-    }, [currentId, pause])
-
-    useEffect(() => {
-        setCount(0)
-    }, [currentId, stories])
+    const { width, height, stories, loop, currentIndex, isPaused } = useContext<GlobalCtx>(GlobalContext)
 
     useEffect(() => {
         if (typeof currentIndex === 'number') {
@@ -46,41 +31,6 @@ export default function () {
         }
     }, [isPaused])
 
-    const incrementCount = () => {
-        setCount((count: number) => {
-            if (count === 0) storyStartCallback()
-            const interval = getCurrentInterval()
-            if (count < 100) {
-                animationFrameId.current = requestAnimationFrame(incrementCount)
-            } else {
-                storyEndCallback()
-                if (currentId === stories.length - 1) {
-                    allStoriesEndCallback()
-                }
-                next()
-            }
-            return count + (100 / ((interval / 1000) * 60))
-        })
-    }
-
-    const storyStartCallback = () => {
-        onStoryStart && onStoryStart(currentId, stories[currentId])
-    }
-
-    const storyEndCallback = () => {
-        onStoryEnd && onStoryEnd(currentId, stories[currentId])
-    }
-
-    const allStoriesEndCallback = () => {
-        onAllStoriesEnd && onAllStoriesEnd()
-    }
-
-    const getCurrentInterval = () => {
-        if (stories[currentId].type === 'video') return videoDuration
-        if (typeof stories[currentId].duration === 'number') return stories[currentId].duration
-        return defaultInterval
-    }
-
     const toggleState = (action: string, bufferAction?: boolean) => {
         setPause(action === 'pause')
         setBufferAction(!!bufferAction)
@@ -89,7 +39,6 @@ export default function () {
     const previous = () => {
         if (currentId > 0) {
             setCurrentId(currentId - 1)
-            setCount(0)
         }
     }
 
@@ -103,13 +52,11 @@ export default function () {
 
     const updateNextStoryIdForLoop = () => {
         setCurrentId((currentId + 1) % stories.length)
-        setCount(0)
     }
 
     const updateNextStoryId = () => {
         if (currentId < stories.length - 1) {
             setCurrentId(currentId + 1)
-            setCount(0)
         }
     }
 
@@ -133,17 +80,14 @@ export default function () {
     const getVideoDuration = (duration: number) => {
         setVideoDuration(duration * 1000)
     }
-
     return (
         <div style={{ ...styles.container, ...{ width, height } }}>
             <ProgressContext.Provider value={{
                 bufferAction: bufferAction,
                 videoDuration: videoDuration,
-                numArray: stories.map((_, i) => i),
-                currentStory: stories[currentId],
                 currentId,
-                count,
-                pause
+                pause,
+                next
             }}>
                 <ProgressArray />
             </ProgressContext.Provider>
