+++ packages/react-insta-stories/src/styles.css.d.ts
@@ -1,7 +1,13 @@
-export interface IStylesCss {
-  "sk-scaleout": string;
-  spinner: string;
+declare namespace StylesCssModule {
+  export interface IStylesCss {
+    "sk-scaleout": string;
+    spinner: string;
+  }
 }
 
-export const locals: IStylesCss;
-export default locals;
+declare const StylesCssModule: StylesCssModule.IStylesCss & {
+  /** WARNING: Only available when `css-loader` is used without `style-loader` or `mini-css-extract-plugin` */
+  locals: StylesCssModule.IStylesCss;
+};
+
+export = StylesCssModule;
